Calysoify: Remove the user meta setting
As a first step towards removing Calypsoify, we are removing the
mechanisms by which the feature is persisted between requests. In
Jetpack this is done by setting a user meta value.

This PR removes that and deletes the user meta if it's present. After
this PR is merged Calypsoify is only activated when the `calypsoify`
query parameter is present.